LayoutTests/imported/w3c/web-platform-tests/html/user-activation/activation-trigger-pointerevent.html

<!DOCTYPE html>
<html>
  <meta charset="utf-8" />
  <title>Test for pointerevent click activation trigger.</title>
  <head>
    <link
      rel="help"
      href="https://html.spec.whatwg.org/multipage/interaction.html#activation-triggering-input-event"
    />
    <meta name="variant" content="?mouse" />
    <meta name="variant" content="?pen" />
    <meta name="variant" content="?touch" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/resources/testdriver.js"></script>
    <script src="/resources/testdriver-actions.js"></script>
    <script src="/resources/testdriver-vendor.js"></script>
    <script src="resources/utils.js"></script>
  </head>
  <body>
    <h1>Just a h1</h1>
  </body>
  <script>
    const pointer_type = location.search.substring(1);
    promise_test(async () => {
      const test_pointer = `${pointer_type}TestPointer`;

      const pointerdown_event = getEvent("pointerdown");
      const pointerup_event = getEvent("pointerup");
      const click_event = getEvent("click");

      const pointer = new test_driver.Actions().addPointer(
        test_pointer,
        pointer_type
      );

      pointer
        .pointerMove(0, 0, {
          origin: document.body,
          sourceName: test_pointer,
        })
        .pointerDown({ sourceName: test_pointer })
        .send();

      await pointerdown_event;
      if (pointer_type === "mouse") {
        assert_true(
          navigator.userActivation.isActive,
          "mouse pointerdown event should result in activation"
        );
        await consumeTransientActivation();
      } else {
        assert_false(
          navigator.userActivation.isActive,
          `${pointer_type} pointerdown event should not result in activation`
        );
      }

      pointer.pointerUp({ sourceName: test_pointer }).send();
      await pointerup_event;

      if (pointer_type === "mouse") {
        assert_false(
          navigator.userActivation.isActive,
          `mouse pointerup should have no activation after pointerdown consumption`
        );
      } else {
        assert_true(
          navigator.userActivation.isActive,
          `${pointer_type} pointerup event should result in activation`
        );
        await consumeTransientActivation();
      }

      await click_event;
      if (pointer_type === "mouse") {
        assert_false(
          navigator.userActivation.isActive,
          `mouse click should have no activation after pointerdown consumption`
        );
      } else {
        assert_false(
          navigator.userActivation.isActive,
          `${pointer_type} click should have no activation after pointerup consumption`
        );
      }
    }, `Activation through ${pointer_type} pointerevent`);
  </script>
</html>
